requirements.txt
fastapi
uvicorn
requests
pyyaml
paho-mqtt